# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8011
info:
  contact:
    email: support@callified.ai
    name: Callified Support
  description: AI-powered outbound calling, WhatsApp CRM & receptionist platform.
  termsOfService: http://callified.ai/terms
  title: Callified Backend API
  version: "1.0"
paths:
...
  /api/auth/sso/jwt:
    get:
      description: Accepts a signed JWT from a trusted issuer, resolves or JIT-creates
        the user, and redirects to the frontend with a Callified session token.
      parameters:
      - description: Signed JWT from issuer
        in: query
        name: token
        required: true
        type: string
      - description: 'Redirect path after login (default: /crm)'
        in: query
        name: redirect
        type: string
      responses:
        "302":
          description: Redirect to frontend with session token
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: SSO JWT login
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: **Bearer {JWT token}**. Obtain a token from POST /api/auth/login
      or POST /api/auth/signup.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"